refactor(reposicao): sessão unificada — stepper persistente, rotas canônicas, Cockpit enxuto
Reestrutura completa da sessão de reposição em 3 fases:

Fase 1 — desbloqueio cognitivo:
- Stepper persistente como layout de rota (ReposicaoSessionLayout), visível
  em todas as telas da sessão, não só no Cockpit
- Vocabulário unificado entre sidebar, stepper e títulos de página
- Banner "continuar de onde parou" lendo o último evento de auditoria do dia
- Tela de Confirmação (etapa 5) com resumo do ciclo + timeline

Fase 2 — rotas canônicas:
- Rotas /admin/reposicao/sessao/* para cada etapa, com redirects de legado
  (LegacyCockpitRedirect preserva ?tab=)
- useReposicaoStatus retorna estado rico por etapa; getStepLocks calcula
  bloqueios (etapa 4 trava com pedidos pendentes, etc.)
- Stepper com cadeado + tooltip nas etapas bloqueadas
- Checklists por etapa (EtapaChecklist) com itens dinâmicos e CTA
- Cockpit vira overview puro (EtapasGrid), sem mais hibridismo rota/aba

Fase 3 — refactor estrutural:
- Cockpit de ~1970 → ~330 linhas, virou orquestrador
- Extraídos: CicloHojePanel, HistoricoComChart, SmartAlertsSection,
  MetricsStrip, AuditLogSection, ShortcutsDialog, ColumnConfig, TabFallback
- Utilitários compartilhados em src/lib/reposicao.ts e src/types/reposicao.ts
- Hooks em src/hooks/useReposicaoSessao.ts
- Removidos 16 casts `as any` desnecessários (tabelas têm tipos gerados)
- 24 testes unitários novos para deriveCurrentStep, getStepLocks e
  buildEtapaChecklist

Build, lint (arquivos novos) e 76 testes verdes.
