+++ src/components/reposicao/ProcessoComprasStepper.tsx
@@ -1,44 +1,60 @@
-import { useNavigate } from "react-router-dom";
-import { Lightbulb, SlidersHorizontal, ClipboardCheck, Upload, Send, ChevronRight, Check } from "lucide-react";
+import { useLocation, useNavigate } from "react-router-dom";
+import {
+  Lightbulb,
+  SlidersHorizontal,
+  ClipboardCheck,
+  Upload,
+  CheckCircle2,
+  ChevronRight,
+  Check,
+  Lock,
+} from "lucide-react";
 import { Card } from "@/components/ui/card";
 import { Badge } from "@/components/ui/badge";
 import { Skeleton } from "@/components/ui/skeleton";
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
 import { cn } from "@/lib/utils";
+import type { StepLock } from "@/hooks/useReposicaoSessao";
 
 type Step = {
   label: string;
   icon: React.ComponentType<{ className?: string }>;
-  to?: string;
+  to: string;
   description?: string;
 };
 
+export const REPOSICAO_STEPS: Step[] = [
+  { label: "Mercado", icon: Lightbulb, to: "/admin/reposicao/sessao/mercado" },
+  { label: "Parâmetros", icon: SlidersHorizontal, to: "/admin/reposicao/sessao/parametros" },
+  { label: "Pedidos", icon: ClipboardCheck, to: "/admin/reposicao/sessao/pedidos" },
+  { label: "Aplicação Omie", icon: Upload, to: "/admin/reposicao/sessao/aplicacao" },
+  { label: "Confirmação", icon: CheckCircle2, to: "/admin/reposicao/sessao/confirmacao" },
+];
+
 interface Props {
   currentStep?: number;
-  /** Optional click handler. When provided, overrides default navigation. */
   onStepClick?: (step: number) => void;
-  /** When true, renders a skeleton placeholder instead of the stepper. */
   isLoading?: boolean;
+  locks?: StepLock[];
 }
 
-export function ProcessoComprasStepper({ currentStep = 3, onStepClick, isLoading = false }: Props) {
+export function ProcessoComprasStepper({
+  currentStep = 3,
+  onStepClick,
+  isLoading = false,
+  locks,
+}: Props) {
   const navigate = useNavigate();
-
-  const steps: Step[] = [
-    { label: "Oportunidades", icon: Lightbulb, to: "/admin/reposicao/mercado" },
-    { label: "Aprovar Parâmetros", icon: SlidersHorizontal, to: "/admin/reposicao/parametros" },
-    { label: "Revisar Pedidos", icon: ClipboardCheck },
-    { label: "Aplicar no Omie", icon: Upload, to: "/admin/reposicao/cockpit?tab=aplicaromie" },
-    { label: "Confirmar Envio", icon: Send, description: "Aguarde o status Disparado" },
-  ];
+  const location = useLocation();
 
   if (isLoading) {
     return (
       <Card className="p-4">
         <div className="flex flex-col sm:flex-row items-stretch sm:items-center gap-2">
-          {steps.map((_, idx) => (
+          {REPOSICAO_STEPS.map((_, idx) => (
             <div key={idx} className="flex items-center flex-1 min-w-0 gap-2">
               <Skeleton className="h-14 w-full rounded-md" />
-              {idx < steps.length - 1 && (
+              {idx < REPOSICAO_STEPS.length - 1 && (
                 <ChevronRight className="hidden sm:block h-4 w-4 text-muted-foreground shrink-0" />
               )}
             </div>
@@ -49,37 +65,42 @@ export function ProcessoComprasStepper({ currentStep = 3, onStepClick, isLoading
   }
 
   return (
-    <Card className="p-4">
-      <ol className="flex flex-col sm:flex-row items-stretch sm:items-center gap-2 sm:gap-0">
-        {steps.map((step, idx) => {
-          const stepNum = idx + 1;
-          const isCurrent = stepNum === currentStep;
-          const isDone = stepNum < currentStep;
-          const isFuture = stepNum > currentStep;
-          const Icon = step.icon;
-          const clickable = !!onStepClick || !!step.to;
+    <TooltipProvider delayDuration={150}>
+      <Card className="p-4">
+        <ol className="flex flex-col sm:flex-row items-stretch sm:items-center gap-2 sm:gap-0">
+          {REPOSICAO_STEPS.map((step, idx) => {
+            const stepNum = idx + 1;
+            const isCurrent = stepNum === currentStep;
+            const isDone = stepNum < currentStep;
+            const isFuture = stepNum > currentStep;
+            const lock = locks?.[idx];
+            const isLocked = !isCurrent && !isDone && !!lock?.locked;
+            const Icon = step.icon;
 
-          const handleClick = () => {
-            if (onStepClick) {
-              onStepClick(stepNum);
-            } else if (step.to) {
-              navigate(step.to);
-            }
-          };
+            const handleClick = () => {
+              if (onStepClick) {
+                onStepClick(stepNum);
+                return;
+              }
+              const targetPath = step.to.split("?")[0];
+              const samePage = location.pathname === targetPath;
+              navigate(step.to, { replace: samePage });
+            };
 
-          return (
-            <li key={step.label} className="flex items-center flex-1 min-w-0">
+            const button = (
               <button
                 type="button"
-                disabled={!clickable}
                 onClick={handleClick}
+                aria-current={isCurrent ? "step" : undefined}
+                aria-disabled={isLocked ? "true" : undefined}
                 className={cn(
                   "flex items-center gap-2 rounded-md px-3 py-2 w-full text-left transition-colors",
                   isCurrent && "bg-primary/10 text-primary border border-primary/30",
                   isDone && "bg-emerald-500/5 text-foreground border border-emerald-500/20",
-                  isFuture && "bg-muted/40 text-muted-foreground border border-transparent",
-                  clickable && !isCurrent && "hover:bg-muted hover:text-foreground cursor-pointer",
-                  !clickable && !isCurrent && "cursor-default",
+                  isFuture && !isLocked && "bg-muted/40 text-muted-foreground border border-transparent",
+                  isLocked && "bg-muted/30 text-muted-foreground/70 border border-dashed border-muted-foreground/20 opacity-70",
+                  !isCurrent && !isLocked && "hover:bg-muted hover:text-foreground cursor-pointer",
+                  isLocked && "cursor-not-allowed",
                 )}
               >
                 <div
@@ -87,48 +108,58 @@ export function ProcessoComprasStepper({ currentStep = 3, onStepClick, isLoading
                     "flex h-8 w-8 shrink-0 items-center justify-center rounded-full text-xs font-semibold",
                     isCurrent && "bg-primary text-primary-foreground",
                     isDone && "bg-emerald-500 text-white",
-                    isFuture && "bg-muted text-foreground/70",
+                    isFuture && !isLocked && "bg-muted text-foreground/70",
+                    isLocked && "bg-muted text-muted-foreground/60",
                   )}
                 >
-                  {isDone ? <Check className="h-4 w-4" /> : <Icon className="h-4 w-4" />}
+                  {isDone ? (
+                    <Check className="h-4 w-4" />
+                  ) : isLocked ? (
+                    <Lock className="h-3.5 w-3.5" />
+                  ) : (
+                    <Icon className="h-4 w-4" />
+                  )}
                 </div>
                 <div className="min-w-0 flex-1">
-                  <div className="flex items-center gap-1.5 flex-wrap">
+                  <div className="flex items-center gap-1.5">
                     <span className="text-[10px] uppercase tracking-wide opacity-70">
                       Etapa {stepNum}
                     </span>
                     {isDone && (
                       <Badge className="h-4 px-1.5 text-[9px] bg-emerald-500 hover:bg-emerald-500 text-white border-0">
-                        Concluído
-                      </Badge>
-                    )}
-                    {isCurrent && (
-                      <Badge className="h-4 px-1.5 text-[9px] bg-primary hover:bg-primary text-primary-foreground border-0">
-                        Etapa {stepNum}
-                      </Badge>
-                    )}
-                    {isFuture && (
-                      <Badge variant="secondary" className="h-4 px-1.5 text-[9px]">
-                        Pendente
+                        ok
                       </Badge>
                     )}
                   </div>
                   <div className="text-sm font-medium truncate">{step.label}</div>
-                  {step.description && (
-                    <div className="text-xs text-muted-foreground truncate">
-                      {step.description}
-                    </div>
-                  )}
                 </div>
               </button>
-              {idx < steps.length - 1 && (
-                <ChevronRight className="hidden sm:block h-4 w-4 text-muted-foreground mx-1 shrink-0" />
-              )}
-            </li>
-          );
-        })}
-      </ol>
-    </Card>
+            );
+
+            return (
+              <li key={step.label} className="flex items-center flex-1 min-w-0">
+                {isLocked && lock?.reason ? (
+                  <Tooltip>
+                    <TooltipTrigger asChild>
+                      <span className="w-full">{button}</span>
+                    </TooltipTrigger>
+                    <TooltipContent side="bottom" className="max-w-xs">
+                      <p className="font-medium mb-1">Etapa bloqueada</p>
+                      <p className="text-xs">{lock.reason}</p>
+                    </TooltipContent>
+                  </Tooltip>
+                ) : (
+                  button
+                )}
+                {idx < REPOSICAO_STEPS.length - 1 && (
+                  <ChevronRight className="hidden sm:block h-4 w-4 text-muted-foreground mx-1 shrink-0" />
+                )}
+              </li>
+            );
+          })}
+        </ol>
+      </Card>
+    </TooltipProvider>
   );
 }
