+++ src/components/reposicao/TabFallback.tsx
@@ -1,0 +1,10 @@
+import { Loader2 } from "lucide-react";
+
+export const TabFallback = () => (
+  <div className="flex items-center justify-center py-16 text-muted-foreground">
+    <Loader2 className="h-5 w-5 animate-spin mr-2" />
+    Carregando...
+  </div>
+);
+
+export default TabFallback;
